Декомпозирует редьюсер
