+++ src/reducer/app-state/app-state.test.js
@@ -1,18 +1,13 @@
 import {INITIAL_FILMS_COUNT, ADDITIONAL_FILMS_COUNT, Screen} from "../const.js";
+import {getFilmsFilteredByGenre} from "../../utils/common.js";
 import {GENRES, films} from "../test-mocks.js";
-import {
-  reducer,
-  ActionType,
-  ActionCreator,
-  getFilmsFilteredByGenre
-} from "./reducer.js";
+import {reducer, ActionType, ActionCreator} from "./app-state.js";
 
 it(`Reducer should change current genre by a given value`, () => {
   expect(reducer({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   }, {
@@ -21,8 +16,7 @@ it(`Reducer should change current genre by a given value`, () => {
   })).toEqual({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[1],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   });
@@ -30,8 +24,7 @@ it(`Reducer should change current genre by a given value`, () => {
   expect(reducer({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   }, {
@@ -39,8 +32,7 @@ it(`Reducer should change current genre by a given value`, () => {
   })).toEqual({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   });
@@ -50,37 +42,37 @@ it(`Reducer should return all films if all films filter was chosen`, () => {
   expect(reducer({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   }, {
-    type: ActionType.FILTER_FILMS_BY_GENRE
+    type: ActionType.FILTER_FILMS_BY_GENRE,
+    payload: films
   })).toEqual({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: getFilmsFilteredByGenre(films, GENRES[0]),
+    filteredFilms: films,
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   });
 });
 
 it(`Reducer should return filtered films list`, () => {
+  const filteredFilms = getFilmsFilteredByGenre(films, GENRES[1]);
+
   expect(reducer({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[1],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   }, {
-    type: ActionType.FILTER_FILMS_BY_GENRE
+    type: ActionType.FILTER_FILMS_BY_GENRE,
+    payload: filteredFilms
   })).toEqual({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[1],
-    films,
-    filteredFilms: getFilmsFilteredByGenre(films, GENRES[1]),
+    filteredFilms,
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   });
@@ -90,8 +82,7 @@ it(`Reducer should reset films count to show`, () => {
   expect(reducer({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT + 1,
     isPlayerActive: false
   }, {
@@ -99,28 +90,7 @@ it(`Reducer should reset films count to show`, () => {
   })).toEqual({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: getFilmsFilteredByGenre(films, GENRES[0]),
-    filmsCountToShow: INITIAL_FILMS_COUNT,
-    isPlayerActive: false
-  });
-});
-
-it(`Reducer should reset films count to show`, () => {
-  expect(reducer({
-    activeScreen: Screen.MAIN,
-    currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
-    filmsCountToShow: INITIAL_FILMS_COUNT + 1,
-    isPlayerActive: false
-  }, {
-    type: ActionType.RESET_FILMS_COUNT_TO_SHOW
-  })).toEqual({
-    activeScreen: Screen.MAIN,
-    currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   });
@@ -130,8 +100,7 @@ it(`Reducer should increment films count to show`, () => {
   expect(reducer({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   }, {
@@ -139,30 +108,30 @@ it(`Reducer should increment films count to show`, () => {
   })).toEqual({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT + ADDITIONAL_FILMS_COUNT,
     isPlayerActive: false
   });
 });
 
 // it(`Reducer should set similar films`, () => {
+//   const similarFilms = getSimilarFilms(films, films[0]);
+
 //   expect(reducer({
 //     activeScreen: Screen.MAIN,
 //     currentGenre: GENRES[0],
-//     films,
-//     filteredFilms: films,
+//     filteredFilms: [],
 //     filmsCountToShow: INITIAL_FILMS_COUNT,
 //     isPlayerActive: false
 //   }, {
-//     type: ActionType.SET_SIMILAR_FILMS
+//     type: ActionType.SET_SIMILAR_FILMS,
+//     payload: similarFilms
 //   })).toEqual({
 //     activeScreen: Screen.MAIN,
 //     currentGenre: GENRES[0],
-//     films,
-//     filteredFilms: films,
+//     filteredFilms: [],
 //     filmsCountToShow: INITIAL_FILMS_COUNT,
-//     similarFilms: getSimilarFilms(films, films[0]),
+//     similarFilms,
 //     isPlayerActive: false
 //   });
 // });
@@ -171,8 +140,7 @@ it(`Reducer should change player state`, () => {
   expect(reducer({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: false
   }, {
@@ -180,8 +148,7 @@ it(`Reducer should change player state`, () => {
   })).toEqual({
     activeScreen: Screen.MAIN,
     currentGenre: GENRES[0],
-    films,
-    filteredFilms: films,
+    filteredFilms: [],
     filmsCountToShow: INITIAL_FILMS_COUNT,
     isPlayerActive: true
   });
@@ -196,8 +163,9 @@ describe(`Action creators work correctly`, () => {
   });
 
   it(`Action creator for filtering films returns correct action`, () => {
-    expect(ActionCreator.filterFilmsByGenre()).toEqual({
-      type: ActionType.FILTER_FILMS_BY_GENRE
+    expect(ActionCreator.filterFilmsByGenre(films)).toEqual({
+      type: ActionType.FILTER_FILMS_BY_GENRE,
+      payload: films
     });
   });
 
@@ -214,8 +182,9 @@ describe(`Action creators work correctly`, () => {
   });
 
   it(`Action creator for setting similar films returns correct action`, () => {
-    expect(ActionCreator.setSimilarFilms()).toEqual({
-      type: ActionType.SET_SIMILAR_FILMS
+    expect(ActionCreator.setSimilarFilms(films)).toEqual({
+      type: ActionType.SET_SIMILAR_FILMS,
+      payload: films
     });
   });
