+++ src/reducer/name-space.js
@@ -1,0 +1,6 @@
+const NameSpace = {
+  APP_STATE: `APP_STATE`,
+  DATA: `DATA`
+};
+
+export default NameSpace;
