+++ src/utils/common.js
@@ -1,5 +1,5 @@
 import moment from "moment";
-import {DURATION_INITS, RUN_TIME_FORMAT} from "../const.js";
+import {DURATION_INITS, RUN_TIME_FORMAT, DEFAULT_GENRE} from "../const.js";
 
 export const getRandomIntegerNumber = (min, max) => {
   min = Math.ceil(min);
@@ -45,3 +45,18 @@ export const getRunTimeForPlayer = (duration) => {
 
   return moment.utc(durationInMillisecond).format(RUN_TIME_FORMAT);
 };
+
+export const getFilmsFilteredByGenre = (allFilms, genre) => {
+  if (genre === DEFAULT_GENRE) {
+    return allFilms;
+  }
+
+  return allFilms.filter((film) => film.genre === genre);
+};
+
+// const getSimilarFilms = (allFilms, currentFilm) => {
+//   return (
+//     allFilms.filter((film) => (film !== currentFilm) && (film.genre === currentFilm.genre))
+//       .slice(0, RECOMENDED_FILMS_COUNT)
+//   );
+// };
